requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.0
psycopg2-binary==2.9.9
python-dotenv==1.2.1